trying to start PWM

+++ Core/Src/measurements.c
@@ -59,7 +59,7 @@ void updateDisplay(OptoMetrics metrics) {
 void measure(TIM_HandleTypeDef *htim, uint32_t pwmChannel, uint16_t *adcValues) {
 	OptoMetrics metrics;
 	// set A6 PWM output to 50%
-	uint32_t drive = 80; // this is not percentage!
+	uint16_t drive = 80; // this is not percentage!
 	HAL_TIM_PWM_Start_DMA(htim, pwmChannel, (uint32_t *)drive, 1);
 
 	// delay
